refactor: 로또 생성과 입출력 로직 클래스 분리
