+++ src/main/java/lotto/model/LottoGenerator.java
@@ -1,0 +1,21 @@
+package lotto.model;
+
+import camp.nextstep.edu.missionutils.Randoms;
+
+import java.util.List;
+import java.util.stream.IntStream;
+
+public class LottoGenerator {
+    public Lotto createSingleLotto() {
+        List<Integer> singleLottoNumber = Randoms.pickUniqueNumbersInRange(1, 45, 6);
+        return new Lotto(singleLottoNumber);
+    }
+
+    public List<Lotto> createMultipleLottos(int ticketCount) {
+        List<Lotto> lottos = IntStream.range(0, ticketCount)
+                .mapToObj(i -> createSingleLotto())
+                .toList();
+
+        return lottos;
+    }
+}
