+++ src/main/java/lotto/model/Purchase.java
@@ -1,0 +1,7 @@
+package lotto.model;
+
+public class Purchase {
+    public int convertMoneyToTicket(int amount) {
+        return amount / 1000;
+    }
+}
