+++ src/test/java/lotto/model/PurchaseTest.java
@@ -1,6 +1,5 @@
-package lotto.temp;
+package lotto.model;
 
-import lotto.model.Lotto;
 import lotto.util.CommonIo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -23,20 +22,4 @@ class PurchaseTest {
         assertThat(purchase.convertMoneyToTicket(100000)).isEqualTo(100);
         assertThat(purchase.convertMoneyToTicket(5000)).isEqualTo(5);
     }
-
-    @Test
-    @DisplayName("발행한 로또가 로또의 객체인지 확인하는 테스트")
-    void createLottoInstance(){
-        assertThat(purchase.createSingleLotto()).isInstanceOf(Lotto.class);
-    }
-
-    @Test
-    @DisplayName("구입 숫자만큼 로또를 발행하는지 확인하는 테스트")
-    void issueTicketByCount(){
-        assertThat(purchase.createMultipleLottos(5)).hasSize(5);
-        assertThat(purchase.createMultipleLottos(1)).hasSize(1);
-        assertThat(purchase.createMultipleLottos(100)).hasSize(100);
-    }
-
-
 }
